REFACTOR Endpoint getAll and delete to custom hook
